Implement help screen; Implement states
Implemented a state machine.  Each state now has a init and frame
routine.  The init routine just draws stuff to the screen and sets up
other things specific to that state.  The frame code only handles the
state specific frame code.

There is now generic frame code and a state change routine.  The state
change routine will handle the timing of turning off the screen and
turning it back on.

The main input code has been moved to its own file and moved to the
first state.

The help screen state has been implemented and is the second state.